refactor(msca): split apart the builder defs a bit to make extension easier
